students/byungmoon/westargram/requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
PyMySQL==1.0.2
PyJWT==2.3.0
PyMySQL==1.0.2
bcrypt==3.2.0